refactor: domain, util, view 패키지로 클래스파일 정리
