+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.domain;
 
 import org.assertj.core.api.Condition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.util.RandomUtil;
 
 import static org.assertj.core.api.Assertions.assertThat;
