+++ src/test/java/racingcar/util/RandomUtilTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import org.junit.jupiter.api.RepeatedTest;
